.qwen/templates/api-service/requirements.txt
# Core FastAPI and async support
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Client for external APIs
httpx==0.25.2

# Logging and Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# API Documentation and Validation
slowapi==0.1.9
redis==5.0.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI
faker==20.1.0  # For generating test data

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Container and Deployment
gunicorn==21.2.0

# Qwen AI Integration (if needed)
# openai==1.3.7  # Uncomment if using OpenAI-compatible APIs
# requests==2.31.0  # For custom API integrations